pyproject.toml
[tool.poetry]
name = "cataclop"
version = "0.1.0"
description = ""
authors = ["Mathias Dusautoy <mathias.dusautoy@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7.*"
django-environ = "*"
django = "*"
djangorestframework = "*"
django-mjml = "^0.11.0"
django-cors-headers = "^3.8.0"
django-redis = "*"
django-configurations = "*"
djangorestframework-simplejwt = "*"
django-filter = "*"
django-extensions = "*"
django-cursor-pagination = "*"
mysqlclient = "*"
pytz = "*"
jupyter = "*"
bencode = "*"
pandas = "*"
scikit-learn = "==0.20.2"
scipy = "*"
tables = "*"
matplotlib = "*"
certifi = "*"
fastparquet = "*"
jinja2 = "*"
requests = "*"
pyopenssl = "*"
idna = "*"
sentry-sdk = "==0.5.5"
joblib = "*"
xgboost = "~=0.9"
seaborn = "*"
theano = "*"
scikit-optimize = "*"
gunicorn = "*"
trueskill = "*"
dill = "*"
faker = "*"
factory-boy = "*"
jupyterlab = "*"
tqdm = "*"
uwsgi = "*"
dash = "==1.19.0"
python-dateutil = "*"
celery = "*"
keras = "*"
tensorflow = "==2.4.1"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "*"
django-debug-toolbar-request-history = "*"